requirements.txt
tensorflow
tensorflow_hub
Pillow
streamlit
numpy
